+++ R/core.R
@@ -15,6 +15,14 @@
 #'
 #' @return TRUE otherwise should throw an error
 #'
+#'
+#' @examples
+#' \dontrun{
+#' Change the default value for `mean` from `0` to `100`
+#' update_function_arguments(function_name = 'rnorm', package_name = 'stats', mean = 100)
+#' }
+#'
+#'
 #' @importFrom utils getFromNamespace
 #' @importFrom rlang dots_definitions f_rhs
 #' @export