+++ NEWS.md
@@ -1,3 +1,10 @@
+
+# grrr 0.1.2
+
+* Add `set_sentinel_on_default_arg()` as a convenient method for a user to 
+  adjust the default arguments of a function to print a message whenever the
+  defaults are used.
+
 # grrr 0.1.1
 
 * Added `default_args_to_global_env()` to dump the default arguments from a